% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialmodifier.R
\name{get_BS_energy}
\alias{get_BS_energy}
\title{Calculate the BS feature energy}
\usage{
get_BS_energy(im, mask)
}
\arguments{
\item{im}{An image.}

\item{mask}{(optional) An image used for mask.}
}
\value{
a data frame
}
\description{
Calculate the BS feature energy
}
\examples{
data = get_BS_energy(face)
}
