% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialmodifier.R
\name{gf_decompose_scale}
\alias{gf_decompose_scale}
\title{Scale-space decomposition by the guided filter}
\usage{
gf_decompose_scale(
  im,
  depth = NULL,
  log_epsilon = 1e-04,
  filter_epsilon = 0.01,
  logspace = TRUE
)
}
\arguments{
\item{im}{a grayscale image}

\item{depth}{scale depth}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}

\item{logspace}{If TRUE (default), image processing is done in the log space. If FALSE,
computation is performed without log transformation.}
}
\value{
a list of images
}
\description{
Scale-space decomposition by the guided filter
}
