% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialmodifier.R
\name{gf_reconstruct}
\alias{gf_reconstruct}
\title{Reconstruct the original image from decomposed data}
\usage{
gf_reconstruct(dec, scales, ind, include.residual = TRUE)
}
\arguments{
\item{dec}{decomposed data}

\item{scales}{which spatial scales to use for reconstruction}

\item{ind}{a numeric vector}

\item{include.residual}{either TRUE (default) or FALSE}
}
\value{
an image
}
\description{
Reconstruct the original image from decomposed data
}
\examples{
\dontrun{
im1 = face
dec = gf_decompose(im1)
im2 = gf_reconstruct(dec)
im_diff(im1, im2) # small difference means successful reconstruction
}
}
