% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialmodifier.R
\name{im_load_dir}
\alias{im_load_dir}
\title{Load all images in a directory and return them as a list}
\usage{
im_load_dir(path)
}
\arguments{
\item{path}{a directory path containing images}
}
\value{
a list of images
}
\description{
Load all images in a directory and return them as a list
}
\examples{
\dontrun{
im = im_load_dir( "path/to/image/folder" )
}
}
