% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face}
\alias{face}
\title{A face image.}
\format{
An array with 500 x 450 x 3 dimensions.
Each dimension represents y-coordinate, x-coordinate, and color channel.
}
\usage{
face
}
\description{
A photograph obtained from a free stock photos site.
pexels.com/photo/close-up-photo-of-woman-wearing-floral-headdress-1453700/
}
\examples{
plot(face)
}
\keyword{datasets}
