% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialmodifier.R
\name{gf_decompose_scale}
\alias{gf_decompose_scale}
\title{Scale-space decomposition by the guided filter}
\usage{
gf_decompose_scale(
  im,
  depth = NULL,
  log_epsilon = 1e-04,
  filter_epsilon = 0.01
)
}
\arguments{
\item{im}{a grayscale image}

\item{depth}{scale depth}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}
}
\value{
a list of images
}
\description{
Scale-space decomposition by the guided filter
}
\examples{
\dontrun{
dec = gf_decompose_scale(im_gray(face))
}
}
